+++ src/@types/PopoutModule/index.d.ts
@@ -4,7 +4,7 @@ declare namespace PopoutModule {
   class Popout {
     public poppedOut: Map<number, PopoutState>
     public ID: string
-    public onPopoutClicked(domID: string, app: Application)
+    public onPopoutClicked(domID: string, app: Application): void
   }
 
   class PopoutState {