+++ src/scripts/Module/constants.ts
@@ -1,4 +1,8 @@
 export const MODULE_KEY = 'focus-sheet'
 
-export const SETTING_AUTO_OPEN = 'auto-open'
-export const SETTING_AUTO_OPEN_POPOUT = 'auto-open-popout'
+export const SETTINGS = {
+  AUTO_OPEN: 'auto-open',
+  AUTO_OPEN_POPOUT: 'auto-open-popout',
+  AUTO_OPEN_POSITION_X: 'auto-open-position-x',
+  AUTO_OPEN_POSITION_Y: 'auto-open-position-y',
+}