+++ src/scripts/Modules/Popout/index.ts
@@ -1,3 +1,4 @@
 export * from './focus'
 export * from './getPopout'
 export * from './open'
+export * from './setPosition'