+++ src/scripts/Settings/Settings.ts
@@ -1,9 +1,15 @@
-import {MODULE_KEY, SETTING_AUTO_OPEN, SETTING_AUTO_OPEN_POPOUT} from '../Module/constants'
+import {MODULE_KEY, SETTINGS} from '../Module/constants'
+
+import {parseNumber} from '../@util/parseNumber'
 
 export interface ModuleSettings {
   AutoOpen: {
     Enabled: boolean
     AsPopout: boolean
+    Position: {
+      X?: number
+      Y?: number
+    }
   }
 }
 
@@ -30,8 +36,12 @@ export class Settings implements ModuleSettings {
 
   public Reset(): void {
     this.autoOpen = {
-      Enabled: game.settings.get(MODULE_KEY, SETTING_AUTO_OPEN),
-      AsPopout: game.settings.get(MODULE_KEY, SETTING_AUTO_OPEN_POPOUT),
+      Enabled: game.settings.get(MODULE_KEY, SETTINGS.AUTO_OPEN),
+      AsPopout: game.settings.get(MODULE_KEY, SETTINGS.AUTO_OPEN_POPOUT),
+      Position: {
+        Y: parseNumber(game.settings.get(MODULE_KEY, SETTINGS.AUTO_OPEN_POSITION_Y)),
+        X: parseNumber(game.settings.get(MODULE_KEY, SETTINGS.AUTO_OPEN_POSITION_X)),
+      },
     }
   }
 }