+++ src/scripts/Sheet/open.ts
@@ -1,14 +1,15 @@
-import {Settings} from '../Settings'
+import {ModuleSettings} from '../Settings'
 import {open as openPopout} from '../Modules/Popout'
 import {waitFor} from '../@util/waitFor'
 
-export async function open(sheet: ActorSheet): Promise<void> {
+export async function open(sheet: ActorSheet, settings: ModuleSettings['AutoOpen']): Promise<void> {
   sheet.render(true)
 
-  await waitFor(() => sheet.rendered)
+  await waitFor(() => sheet.rendered, {title: `Rendering sheet for ${sheet.actor.name}`})
 
-  const settings = Settings.GetInstance()
-  if (settings.AutoOpen.AsPopout) {
-    openPopout(sheet)
+  if (settings.AsPopout) {
+    openPopout(sheet, settings.Position)
+  } else {
+    sheet.setPosition({left: settings.Position.X, top: settings.Position.Y})
   }
 }