+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.3.0
+
+### Features
+
+* add main feature: configure sheet position when auto-opening
+
+## v0.2.1
+
+_No changes to the module logic, just some fixed references, e.g. the changelog link._
+
 ## v0.2.0
 
 ### Features